src/requirements.txt
couchbase==4.3.5
langchain-community==0.3.20
langchain-couchbase==0.2.4
langchain-huggingface==0.1.2
langchain==0.3.21
python-dotenv==1.0.1
sentence-transformers==3.4.1